feat(kqueue): implement IoEngine/Channel/ServerChannel
- KqueueEngine: IoEngine with suspend bind/connect (Phase (a) sync)
- KqueueChannel: zero-copy read/write via NativeBuf.unsafePointer,
  write/flush separation with PendingWrite buffer, shutdownOutput
- KqueueServerChannel: kevent-based accept with kqueue readiness wait
- ChannelSource/ChannelSink: kotlinx-io RawSource/RawSink bridge
- Remove legacy runEchoLoop/acceptAndRegister/echoOnce API

+++ engine-kqueue/src/macosMain/kotlin/io/github/keel/engine/kqueue/SocketUtils.kt
@@ -56,7 +56,7 @@ internal object SocketUtils {
             addr.sin_family = AF_INET.convert()
             addr.sin_port = keel_htons(port.toUShort())
             if (host == "0.0.0.0") {
-                addr.sin_addr.s_addr = INADDR_ANY.toUInt()
+                addr.sin_addr.s_addr = INADDR_ANY
             } else {
                 val rc = inet_pton(AF_INET, host, addr.sin_addr.ptr)
                 check(rc == 1) { "Invalid address: $host" }
